chore(examples): move examples from yarn to pnpm
