+++ examples/misc/ckb-indexer-collector.ts
