+++ examples/misc/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@lumos-example/misc",
+  "private": true,
+  "version": "1.0.0",
+  "license": "MIT",
+  "dependencies": {
+    "@ckb-lumos/lumos": "0.20.0-alpha.3"
+  }
+}
