+++ examples/omni-lock-metamask/package.json
@@ -10,10 +10,25 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@ckb-lumos/base": "0.20.0-alpha.3",
+    "@ckb-lumos/codec": "0.20.0-alpha.3",
     "@ckb-lumos/lumos": "0.20.0-alpha.3",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.9",
     "react": "18.2.0",
     "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "parcel": "^2.9.3",
+    "buffer": "^5.5.0",
+    "crypto-browserify": "^3.12.0",
+    "events": "^3.1.0",
+    "path-browserify": "^1.0.0",
+    "process": "^0.11.10",
+    "stream-browserify": "^3.0.0",
+    "string_decoder": "^1.3.0"
+  },
+  "alias": {
+    "process": false
   }
 }
