+++ examples/omni-lock-secp256k1-blake160/package.json
@@ -10,10 +10,17 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@ckb-lumos/base": "0.20.0-alpha.3",
+    "@ckb-lumos/codec": "0.20.0-alpha.3",
+    "@ckb-lumos/hd": "0.20.0-alpha.3",
     "@ckb-lumos/lumos": "0.20.0-alpha.3",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.9",
+    "elliptic": "^6.5.4",
     "react": "18.2.0",
     "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "parcel": "^2.9.3"
   }
 }
