+++ examples/package.json
@@ -3,19 +3,13 @@
   "private": true,
   "version": "1.0.0",
   "license": "MIT",
-  "workspaces": [
-    "../packages/*",
-    "cardano-lock-namiwallet",
-    "omni-lock-metamask",
-    "omni-lock-secp256k1-blake160",
-    "pw-lock-metamask",
-    "secp256k1-transfer",
-    "exchange-sudt-for-ckb",
-    "secp256k1-multisig-transfer"
-  ],
   "devDependencies": {
+    "@parcel/config-default": "^2.8.0",
+    "@parcel/transformer-js": "^2.8.0",
+    "@parcel/transformer-react-refresh-wrap": "^2.8.0",
     "buffer": "^5.5.0",
-    "parcel": "2.8.0",
+    "crypto-browserify": "^3.12.0",
+    "events": "^3.1.0",
     "path-browserify": "^1.0.0",
     "stream-browserify": "^3.0.0",
     "string_decoder": "^1.3.0"
@@ -23,7 +17,5 @@
   "alias": {
     "process": false,
     "stream": false
-  },
-  "dependencies": {
   }
 }
