+++ examples/secp256k1-multisig-transfer/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@lumos-examples/secp256k1-transfer",
+  "name": "@lumos-examples/secp256k1-multisig-transfer",
   "version": "0.20.0-alpha.3",
   "description": "",
   "main": "index.js",
@@ -13,14 +13,9 @@
   "license": "MIT",
   "dependencies": {
     "@ckb-lumos/base": "0.20.0-alpha.3",
+    "@ckb-lumos/bi": "0.20.0-alpha.3",
     "@ckb-lumos/codec": "0.20.0-alpha.3",
-    "@ckb-lumos/lumos": "0.20.0-alpha.3",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.9",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "parcel": "^2.9.3"
+    "@ckb-lumos/common-scripts": "0.20.0-alpha.3",
+    "@ckb-lumos/lumos": "0.20.0-alpha.3"
   }
 }
